[YS-151] refactor: 회원가입 이메일 인증 화면 컴포넌트로 분리
